+++ ukrdc_fastapi/routers/api/__init__.py
@@ -1,13 +1,10 @@
 from fastapi import APIRouter
 from starlette.responses import RedirectResponse
 
-from ukrdc_fastapi.config import settings
-
 from . import (
     admin,
     codes,
     dashboard,
-    debug,
     empi,
     facilities,
     masterrecords,
@@ -55,7 +52,3 @@ router.include_router(messages.router, prefix="/messages")
 
 # Task management
 router.include_router(tasks.router, prefix="/tasks")
-
-# Debug routes
-if settings.debug:
-    router.include_router(debug.router, prefix="/debug")
